+++ lib/src/call_invitation/pages/calling_page.dart
@@ -151,6 +151,7 @@ class ZegoCallingPageState extends State<ZegoCallingPage> {
         widget.pageManager.onPrebuiltCallPageDispose();
       },
       controller: widget.callInvitationConfig.controller,
+      plugins: widget.callInvitationConfig.plugins,
     );
   }
 }